Normalize integral promotion boundaries
Remove redundant word-to-int and same-type promotion ladders while preserving listing-proven narrowing, packed-field, and signedness boundaries. Correct amount-bar geometry to use full-width TView dimensions and recover the short ApplyMoveClamp virtual return type.

Verified with just agent-check and just precommit: all vtables 100%, 452 tests passed with 4 skipped, exact functions +2, and average similarity +0.01 pp. Refreshed reccmp progress baselines.

+++ src/game/TCivDescription.cpp
@@ -228,7 +228,7 @@ void TCivDescription::UpdateCivilianOrderTargetTileCountsForOwnerNation(TCivUnit
       do {
         provinceTileIndex = (short)*provinceTileIndices;
         tableBase = reinterpret_cast<char*>(g_pGlobalMapState->terrainStateTable) +
-                    (int)(short)provinceTileIndex * 0x24;
+                    static_cast<short>(provinceTileIndex) * 0x24;
         if (*(char*)(tableBase + 0xe) == '\0') {
           tileProfileId = (short)*(char*)(tableBase + 0x13);
           classSlotOrdinal = 0;
