+++ src/game/TMacViewMgr.cpp
@@ -1603,7 +1603,7 @@ undefined TMacViewMgr::BlitMapOverlayGlyphStrip32x24SkipMask10(TBitmapSurfaceNod
   unsigned char* srcPixels = GetPixBaseAddr(atlasSurface);
   ushort srcStrideRaw = static_cast<ushort>((*atlasSurface)->stride);
   unsigned char* dstPixels = GetPixBaseAddr(dstSurface);
-  int dstStrideBytes = static_cast<int>(static_cast<short>(dstStrideRaw & 0x3fff));
+  int dstStrideBytes = static_cast<short>(dstStrideRaw & 0x3fff);
   unsigned char* srcRow = srcPixels + srcRowOffset;
   unsigned char* dstRow = dstPixels + param_4 * dstStrideBytes + param_3;
   int rowsRemaining = 0x18;
@@ -1689,8 +1689,7 @@ void TMacViewMgr::DrawStrategicMapUnitIcon(TBitmapSurfaceNode** pDstSurface, sho
   unsigned char* srcPixels = GetPixBaseAddr(atlasSurface);
   ushort srcStrideRaw = static_cast<ushort>((*atlasSurface)->stride);
   unsigned char* dstPixels = GetPixBaseAddr(pDstSurface);
-  int dstStrideBytes =
-      static_cast<int>(static_cast<short>(static_cast<ushort>((*pDstSurface)->stride) & 0x3fff));
+  int dstStrideBytes = static_cast<short>(static_cast<ushort>((*pDstSurface)->stride) & 0x3fff);
   unsigned char* srcRow = srcPixels + static_cast<short>(nIconVariant * 0x14);
   unsigned char* dstRow = dstPixels + (0x28 - nYShift) * dstStrideBytes + static_cast<int>(nDstX);
   int rowsRemaining = 0x18;
@@ -1759,8 +1758,7 @@ void TMacViewMgr::DrawStrategicMapUnitIconOverlay(TBitmapSurfaceNode** pDstSurfa
   unsigned char* srcPixels = GetPixBaseAddr(atlasSurface);
   ushort srcStrideRaw = static_cast<ushort>((*atlasSurface)->stride);
   unsigned char* dstPixels = GetPixBaseAddr(pDstSurface);
-  int dstStrideBytes =
-      static_cast<int>(static_cast<short>(static_cast<ushort>((*pDstSurface)->stride) & 0x3fff));
+  int dstStrideBytes = static_cast<short>(static_cast<ushort>((*pDstSurface)->stride) & 0x3fff);
   unsigned char* srcRow =
       srcPixels + static_cast<short>(overlaySourceOffset - 0x26 + nVariantRow * 0x26);
   unsigned char* dstRow = dstPixels + (0x28 - nYShift) * dstStrideBytes + static_cast<int>(nDstX);
